refactor(state): split intent logic out of StateCoordinator
The sticky intent memory (#414) pushed StateCoordinator past
the 350-line ceiling. Move the manual float/sticky intent
logic -- setFloating / setSticky / clearFloatOverride and the
remember/restore close-reopen memory -- into
StateCoordinator+Intents.swift. The stored maps stay on the
struct; windows becomes internal(set) (like workspaces) so the
split file can mutate it.
